feat: Prepare the HomeListingsVC to present data to the view
A collection of homes is provided by the managed object context. The
managed object context is being initialized in the app delegate.  The
data is now available to be presented in the table view.

To display the data, the custom table view cell will need to include more
outlets for the UI components.